+++ src/app/controller/common.go
@@ -8,6 +8,9 @@ import (
 	"github.com/davyzhang/agw"
 )
 
+const GRAPHQL_API = "https://api.github.com/graphql"
+const LABEL_PREVIEW_HEADER = "application/vnd.github.bane-preview+json"
+
 func sendJson(w http.ResponseWriter, status int, payload interface{}) {
 	response, err := json.Marshal(payload)
 	if err != nil {